backend/africa_contech_hub/requirements.txt
Django>=5.0
python-dotenv
psycopg2-binary
dj-database-url
djangorestframework
supabase
django-cors-headers
